fix(ai): isolate provider tool execution to server utilities
