+++ apps/tradinggoose/providers/ai/ollama/index.ts
@@ -10,11 +10,11 @@ import type {
   TimeSegment,
 } from '@/providers/ai/types'
 import {
-  executeProviderTool,
   prepareToolExecution,
   prepareToolsWithUsageControl,
   trackForcedToolUsage,
 } from '@/providers/ai/utils'
+import { executeProviderTool } from '@/providers/ai/utils-server'
 import { useProvidersStore } from '@/stores/providers/store'
 
 const logger = createLogger('OllamaProvider')
