+++ apps/tradinggoose/providers/ai/utils-server.ts
@@ -1,5 +1,54 @@
 import { isHosted } from '@/lib/environment'
+import type { ProviderRequest } from '@/providers/ai/types'
 import { useProvidersStore } from '@/stores/providers/store'
+import { executeTool } from '@/tools'
+import type { ToolResponse } from '@/tools/types'
+
+export async function executeProviderTool(
+  request: ProviderRequest,
+  toolId: string,
+  params: Record<string, any>,
+  skipPostProcess = false
+): Promise<ToolResponse> {
+  const operation = await request.beginToolOperation?.(toolId)
+  let identityPublished = false
+  let terminalObserved = false
+  const runtime = operation
+    ? {
+        ...operation.runtime,
+        publishOperationIdentity: async (
+          identity: Parameters<NonNullable<typeof operation.runtime.publishOperationIdentity>>[0]
+        ) => {
+          identityPublished = true
+          await operation.runtime.publishOperationIdentity?.(identity)
+        },
+        recordTerminalObservation: async (
+          state: 'canceled' | 'completed' | 'failed',
+          observation?: Record<string, unknown>
+        ) => {
+          terminalObserved = true
+          await operation.runtime.recordTerminalObservation?.(state, observation)
+        },
+      }
+    : { signal: request.abortSignal }
+
+  try {
+    const result = await executeTool(toolId, params, skipPostProcess, undefined, runtime)
+    if (operation && !terminalObserved && !identityPublished) {
+      await operation.finish(result.success ? 'completed' : 'failed')
+    }
+    return result
+  } catch (error) {
+    if (operation && !terminalObserved) {
+      if (request.abortSignal?.aborted) {
+        await operation.finish('local_abort')
+      } else if (!identityPublished) {
+        await operation.finish('failed')
+      }
+    }
+    throw error
+  }
+}
 
 /**
  * Get an API key for a specific provider, handling rotation and fallbacks.
